+++ package.json
@@ -11,7 +11,7 @@
   "scripts": {
     "analyze": "bash ./scripts/analyze.sh",
     "build": "bash ./scripts/build.sh",
-    "dev": "bash ./scripts/dev.sh",
+    "dev": "bash ./scripts/dev.sh chrome",
     "lint": "concurrently \"lint:*\"",
     "lint:fix": "eslint --ext js,jsx,ts,tsx, src --fix",
     "lint:eslint": "eslint . --ext .js,.ts,.tsx --max-warnings 0 --ignore-path .gitignore",
