+++ src/manifest/chrome-v3.json
@@ -1,39 +1,37 @@
 {
-    "name": "__MSG_appName__",
-    "version": "1.0.1",
-    "manifest_version": 3,
-    "description": "__MSG_appDescription__",
-    "icons": {
-        "34": "assets/icons/icon-34.png",
-        "128": "assets/icons/icon-128.png"
+  "name": "__MSG_appName__",
+  "version": "1.0.1",
+  "manifest_version": 3,
+  "description": "__MSG_appDescription__",
+  "icons": {
+    "34": "assets/icons/icon-34.png",
+    "128": "assets/icons/icon-128.png"
+  },
+  "default_locale": "en",
+  "content_scripts": [
+    {
+      "matches": ["http://*/*", "https://*/*"],
+      "js": ["content/content.js"]
     },
-    "default_locale": "en",
-    "content_scripts": [
-        {
-            "matches": ["http://*/*", "https://*/*"],
-            "js": ["content/content.js"]
-        }
-    ],
-    "background": {
-        "service_worker": "background/background.js"
-    },
-    "permissions": ["tabs", "storage"],
-    "host_permissions": ["http://*/*", "https://*/*"],
-    "options_ui": {
-        "page": "options/index.html"
-    },
-    "action": {
-        "default_icon": {
-            "16": "assets/icons/icon-16.png",
-            "48": "assets/icons/icon-48.png"
-        },
-        "default_title": "Web Monetization",
-        "default_popup": "popup/index.html"
-    },
-    "web_accessible_resources": [
-        {
-            "resources": ["assets/*", "content/*", "options/*", "popup/*", "background/*"],
-            "matches": ["<all_urls>"]
-        }
-    ]
+    {
+      "run_at": "document_start",
+      "matches": ["http://*/*", "https://*/*"],
+      "js": ["contentStatic/contentStatic.js"]
+    }
+  ],
+  "background": {
+    "service_worker": "background/background.js"
+  },
+  "permissions": ["tabs", "storage"],
+  "host_permissions": ["http://*/*", "https://*/*"],
+  "action": {
+    "default_title": "Web Monetization",
+    "default_popup": "popup/index.html"
+  },
+  "web_accessible_resources": [
+    {
+      "resources": ["assets/*", "content/*", "options/*", "popup/*", "background/*"],
+      "matches": ["<all_urls>"]
+    }
+  ]
 }
